+++ src/index.jsx
@@ -11,6 +11,15 @@ const store = setupStore({});
 const container = document.getElementById("root") || document.createElement("div");
 const root = createRoot(container);
 
+/* istanbul ignore next */ 
+if (process.env.NODE_ENV === "development" && process.env.REACT_APP_MSW) {
+  /* istanbul ignore next */ 
+  // eslint-disable-next-line global-require
+  const { worker } = require("./__tests__/__mocks__/api/browser");
+  /* istanbul ignore next */ 
+  worker.start();
+}
+
 root.render(
   <BrowserRouter>
     <React.StrictMode>