+++ src/__tests__/App.test.js
@@ -4,33 +4,14 @@ import { renderWithProviders } from "./__test-utils__/test-utils";
 import { BrowserRouter } from "react-router-dom";
 import { email as mockedEmail } from "./__mocks__/items/email";
 import { password as mockedPassword } from "./__mocks__/items/password";
-import { item as loginResponse } from "./__mocks__/responses/login";
-import { item as balanceResponse } from "./__mocks__/responses/balance";
 import { rest } from "msw";
 // eslint-disable-next-line jest/no-mocks-import
 import { server } from "./__mocks__/api/server";
 
-beforeEach(() => {
-  server.use(
-    rest.post(
-      `${process.env.REACT_APP_BACKEND_URL}/auth/login`,
-      (req, res, ctx) => {
-        return res(ctx.json(loginResponse));
-      }
-    )
-  );
-  server.use(
-    rest.get(`${process.env.REACT_APP_BACKEND_URL}/balance`, (req, res, ctx) => {
-      return res(ctx.json(balanceResponse));
-    })
-  );
-});
+beforeEach(() => { server.listen(); });
 // Reset any request handlers that we may add during the tests,
 // so they don't affect other tests.
-afterEach(() => {
-  server.resetHandlers();
-});
-
+afterEach(() => { server.resetHandlers(); });
 // Clean up after the tests are finished.
 afterAll(() => server.close());