+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 + Add MSW during development
 + Refactor MSW tests
 + Refactor tests
++ Upgrade dependencies
 
 ## 0.11.0
 + Upgrade dependencies + Refactor Semaphore